Split tests into two batches to work around timeout issues.
