+++ NetMQ.ReactiveExtensions.Tests/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
+
 <packages>
   <package id="AsyncIO" version="0.1.26.0" targetFramework="net40" />
   <package id="NetMQ" version="3.3.3.4" targetFramework="net40" />
